+++ authomatic/providers/oauth2.py
@@ -29,8 +29,10 @@ Providers which implement the |oauth2|_ protocol.
     
 """
 
-from authomatic.six.moves.urllib.parse import urlencode
+from authomatic.six.moves.urllib.parse import urlencode, unquote
+import base64
 import datetime
+import json
 import logging
 
 from authomatic import providers
@@ -58,6 +60,9 @@ class OAuth2(providers.AuthorizationProvider):
     
     #: :class:`bool` If ``False``, the provider doesn't support CSRF protection.
     supports_csrf_protection = True
+    
+    #: :class:`bool` If ``False``, the provider doesn't support user_state.
+    supports_user_state = True
     
     token_request_method = 'POST'  # method for requesting an access token
     
@@ -100,7 +105,8 @@ class OAuth2(providers.AuthorizationProvider):
     
     @classmethod
     def create_request_elements(cls, request_type, credentials, url, method='GET', params=None,
-                                headers=None, body='', secret=None, redirect_uri='', scope='', csrf=''):
+                                headers=None, body='', secret=None, redirect_uri='', scope='',
+                                csrf='', user_state=''):
         """
         Creates |oauth2| request elements.
         """
@@ -126,7 +132,10 @@ class OAuth2(providers.AuthorizationProvider):
                 params['client_id'] = consumer_key
                 params['redirect_uri'] = redirect_uri
                 params['scope'] = scope
-                params['state'] = csrf
+                if cls.supports_user_state:
+                    params['state'] = base64.urlsafe_b64encode(json.dumps({"csrf": csrf, "user_state": user_state}).encode('utf-8'))
+                else:
+                    params['state'] = csrf
                 params['response_type'] = 'code'
                 
                 # Add authorization header
@@ -221,9 +230,32 @@ class OAuth2(providers.AuthorizationProvider):
         credentials.token = token
         credentials.refresh_token = refresh_token
         credentials.expiration_time = expiration_time
-        credentials.token_type=cls.TOKEN_TYPES[int(token_type)]
+        credentials.token_type = cls.TOKEN_TYPES[int(token_type)]
         
         return credentials
+    
+    
+    @classmethod
+    def decode_state(cls, state, param='user_state'):
+        """Decode state and return param
+
+        :param str state:
+            state parameter passed through by provider
+        
+        :param str param:
+            key to query from decoded state variable. Options include 'csrf'
+            and 'user_state'.  
+        
+        :returns:
+            string value from decoded state
+        """
+        if state and cls.supports_user_state:
+            # urlsafe_b64 may include = which the browser quotes so must unquote
+            # Cast to str to void b64decode translation error. Base64 should be
+            # str compatible.
+            return json.loads(base64.urlsafe_b64decode(unquote(str(state))).decode('utf-8'))[param]
+        else:
+            return state if param == 'csrf' else ''
     
     
     def refresh_credentials(self, credentials):
@@ -283,7 +315,9 @@ class OAuth2(providers.AuthorizationProvider):
         authorization_code = self.params.get('code')
         error = self.params.get('error')
         error_message = self.params.get('error_message')
-        state = self.params.get('state')      
+        state = self.params.get('state')
+        # optional user_state to be passed in oauth2 state
+        user_state = self.params.get('user_state', '')
         
         if authorization_code or not self.user_authorization_url:
             
@@ -297,12 +331,13 @@ class OAuth2(providers.AuthorizationProvider):
                 # validate CSRF token
                 if self.supports_csrf_protection:
                     self._log(logging.INFO, u'Validating request by comparing request state with stored state.')
-                    stored_state = self._session_get('state')
+                    stored_csrf = self._session_get('csrf')
                     
-                    if not stored_state:
+                    state_csrf = self.decode_state(state, 'csrf')
+                    if not stored_csrf:
                         raise FailureError(u'Unable to retrieve stored state!')
-                    elif not stored_state == state:
-                        raise FailureError(u'The returned state "{0}" doesn\'t match with the stored state!'.format(state),
+                    elif not stored_csrf == state_csrf:
+                        raise FailureError(u'The returned state csrf cookie "{0}" doesn\'t match with the stored state!'.format(state_csrf),
                                            url=self.user_authorization_url)
                     self._log(logging.INFO, u'Request is valid.')
                 else:
@@ -381,7 +416,7 @@ class OAuth2(providers.AuthorizationProvider):
             else:
                 raise FailureError(error_description, url=self.user_authorization_url)
             
-        elif not self.params:
+        elif not self.params or len(self.params) == 1 and 'user_state' in self.params:
             #===================================================================
             # Phase 1 before redirect
             #===================================================================
@@ -393,7 +428,7 @@ class OAuth2(providers.AuthorizationProvider):
                 # generate csfr
                 csrf = self.csrf_generator(self.settings.secret)
                 # and store it to session
-                self._session_set('state', csrf)
+                self._session_set('csrf', csrf)
             else:
                 self._log(logging.WARN, u'Provider doesn\'t support CSRF validation!')
                         
@@ -403,6 +438,7 @@ class OAuth2(providers.AuthorizationProvider):
                                                             redirect_uri=self.url,
                                                             scope=self._x_scope_parser(self.scope),
                                                             csrf=csrf,
+                                                            user_state=user_state,
                                                             params=self.user_authorization_params)
             
             self._log(logging.INFO, u'Redirecting user to {0}.'.format(request_elements.full_url))
@@ -579,7 +615,7 @@ class Bitly(OAuth2):
         super(Bitly, self).__init__(*args, **kwargs)
         
         if self.offline:
-            if not 'grant_type' in self.access_token_params:
+            if 'grant_type' not in self.access_token_params:
                 self.access_token_params['grant_type'] = 'refresh_token'
     
     @staticmethod
@@ -667,7 +703,7 @@ class DeviantART(OAuth2):
         super(DeviantART, self).__init__(*args, **kwargs)
         
         if self.offline:
-            if not 'grant_type' in self.access_token_params:
+            if 'grant_type' not in self.access_token_params:
                 self.access_token_params['grant_type'] = 'refresh_token'
     
     
@@ -833,8 +869,8 @@ class Facebook(OAuth2):
         _birth_date = data.get('birthday')
         if _birth_date:
             try:
-              user.birth_date = datetime.datetime.strptime(_birth_date,
-                                                           '%m/%d/%Y')
+                user.birth_date = datetime.datetime.strptime(_birth_date,
+                                                             '%m/%d/%Y')
             except ValueError:
                 pass
 
@@ -859,7 +895,10 @@ class Facebook(OAuth2):
         
         # Facebook returns "expires" instead of "expires_in".
         credentials.expire_in = data.get('expires')
-        
+
+        if data.get('token_type') == 'bearer':
+            credentials.token_type = cls.BEARER
+
         return credentials
     
     
@@ -867,6 +906,14 @@ class Facebook(OAuth2):
     def _x_refresh_credentials_if(credentials):
         # Always refresh.
         return True
+
+
+    def access(self, url, params=None, **kwargs):
+        if params is None:
+            params = {}
+        params['fields'] = 'id,first_name,last_name,picture,email,gender,timezone,location,birthday,locale'
+
+        return super(Facebook, self).access(url, params, **kwargs)
 
 
 class Foursquare(OAuth2):
@@ -1116,7 +1163,7 @@ class Google(OAuth2):
     
     user_authorization_url = 'https://accounts.google.com/o/oauth2/auth'
     access_token_url = 'https://accounts.google.com/o/oauth2/token'
-    user_info_url = 'https://www.googleapis.com/plus/v1/people/me'
+    user_info_url = 'https://www.googleapis.com/oauth2/v3/userinfo?alt=json'
 
     user_info_scope = ['profile',
                        'email']
@@ -1127,9 +1174,7 @@ class Google(OAuth2):
         name=True,
         first_name=True,
         last_name=True,
-        gender=True,
         locale=True,
-        link=True,
         picture=True
     )
     
@@ -1138,10 +1183,10 @@ class Google(OAuth2):
         
         # Handle special Google requirements to be able to refresh the access token.
         if self.offline:
-            if not 'access_type' in self.user_authorization_params:
+            if 'access_type' not in self.user_authorization_params:
                 # Google needs access_type=offline param in the user authorization request.
                 self.user_authorization_params['access_type'] = 'offline'
-            if not 'approval_prompt' in self.user_authorization_params:
+            if 'approval_prompt' not in self.user_authorization_params:
                 # And also approval_prompt=force.
                 self.user_authorization_params['approval_prompt'] = 'force'
 
@@ -1169,17 +1214,15 @@ class Google(OAuth2):
                     user.email = email.get('value')
                     break
 
-        user.id = data.get('sub') or data.get('id')
-        user.name = data.get('displayName')
-        user.first_name = data.get('name',{}).get('givenName')
-        user.last_name = data.get('name',{}).get('familyName')
-        user.locale = data.get('language')
-        user.link = data.get('url')
-        user.picture = data.get('image',{}).get('url')
-        try:
-            user.birth_date = datetime.datetime.strptime(data.get('birthdate'), "%Y-%m-%d")
-        except:
-            user.birth_date = data.get('birthdate')
+        user.id = data.get('sub')
+        user.name = data.get('name')
+        user.first_name = data.get('given_name', '')
+        user.last_name = data.get('family_name', '')
+        user.locale = data.get('locale', '')
+        user.picture = data.get('picture', '')
+
+        user.email_verified = data.get("email_verified")
+        user.hosted_domain = data.get("hd")
         return user
     
     def _x_scope_parser(self, scope):
@@ -1399,7 +1442,7 @@ class Reddit(OAuth2):
         super(Reddit, self).__init__(*args, **kwargs)
         
         if self.offline:
-            if not 'duration' in self.user_authorization_params:
+            if 'duration' not in self.user_authorization_params:
                 # http://www.reddit.com/r/changelog/comments/11jab9/reddit_change_permanent_oauth_grants_using/
                 self.user_authorization_params['duration'] = 'permanent'
 
@@ -1548,7 +1591,7 @@ class VK(OAuth2):
         super(VK, self).__init__(*args, **kwargs)
         
         if self.offline:
-            if not 'offline' in self.scope:
+            if 'offline' not in self.scope:
                 self.scope.append('offline')
     
     
@@ -1627,7 +1670,7 @@ class WindowsLive(OAuth2):
         super(WindowsLive, self).__init__(*args, **kwargs)
         
         if self.offline:
-            if not 'wl.offline_access' in self.scope:
+            if 'wl.offline_access' not in self.scope:
                 self.scope.append('wl.offline_access')
     
     @classmethod